CssTransitionEvent only takes in target style
I wanted to remove some redundancy here and since we already have a
SetStyleEvent one could use that to set the initial style and
CssTransitionEvent then is used only for the animation to the next state.
This seems good especially if I start adding JS versions of the CSS
animations since it won't be duplicating that logic.